requirements.txt
blinker
certifi
charset-normalizer
click
cloudpickle
colorama
dash
dash-core-components
dash-html-components
dash-table
dash_daq
dask
Faker
Flask
fsspec
idna
importlib_metadata
itsdangerous
Jinja2
kafka-python
locket
MarkupSafe
mysql-connector-python
nest-asyncio
numpy
packaging
pandas
partd
plotly
py4j
pyspark
python-dateutil
pytz
PyYAML
requests
retrying
setuptools
six
tenacity
toolz
typing_extensions
tzdata
urllib3
Werkzeug
zipp
